requirements.txt
torch
locache==2.1
numpy
pandas
scikit-learn
scipy
scikit-optimize
ipykernel
matplotlib
imageio
quippy-ase
ase
pytorch-lightning
